Javascript and CSS styling for dynamic table
Splits the JavaScript into two files - one to handle form validation and tooltips, and the other to handle asynchronous GET request to NREL Electricity Incentives API and dynamic creation of table in results.html page. Uses XMLHttpRequest object and callback function that is bound to onreadystatechange property of XMLHttpRequest to manage the asynchronous request.

CSS file adds styling for table, as well as a loader that runs until the async GET request and JSON parsing is finished.