Check if a zip code is real
Before this change, the application was only checking if a zip code was 5 digits, not if it was a real zip code. I obtained a list of all valid US zip codes (zipCodesNormalized.txt) file, and form validation now checks against this list, once 5 digits have been entered. Form cannot be submitted if the zip code is not real.